services/requirements.txt
# pip3 install -r requirements.txt
flask == 2.1.1
werkzeug == 2.3.6
flask_cors == 3.0.9
psycopg2-binary == 2.9.6
dotted-dict == 1.1.3
rich == 13.3.5
pytest-xdist == 3.3.0
pytest == 7.3.1
coverage == 7.2.5
python-docx == 0.8.11
docxtpl == 0.16.7
joblib == 1.2.0
nltk == 3.8.1
bs4 == 0.0.1
PyJWT == 2.4.0
Pillow == 9.5.0
litellm==1.7.7
stytch==8.0.4
json_repair==0.4.5
apprise==1.7.2
sendgrid==6.11.0
tabulate==0.9.0
whoosh==2.7.4
python-magic==0.4.27
stytch==8.0.4
apache-libcloud==3.8.0

# for langchain use below command :- 
# pip3 install --upgrade langchain
